updates on driver dashboard and its menus

+++ src/main/java/com/group13/fleet/repository/DriverRepository.java
@@ -7,7 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface DriverRepository extends JpaRepository<Driver, Long> {
+public interface DriverRepository extends JpaRepository<Driver, Integer> {
     Optional<Driver> findByUsername(String username);
     Driver findByEmail(String email);
+
 }
